backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
yt-dlp
ffmpeg-python
python-dotenv
openai
deepgram-sdk
celery
redis
sqlalchemy
pydantic-settings
